// src/shared/enums/buildings.enum.ts
export enum EBuildings {
  BAKERY = "bakery",
  BARN = "barn",
  BOOKSTORE = "bookstore",
  BREWERY = "brewery",
  CARPENTER = "carpenter",
  CASTLE = "castle",
  CATHEDRAL = "cathedral",
  CHURCH = "church",
  FOUNTAIN = "fountain",
  FISHERMAN = "fisherman",
  GOLDMINE = "gold mine",
  LOMBARD = "lombard",
  MARKET = "market",
  PIER = "pier",
  QUARRY = "quarry",
  RESTAURANT = "restaurant",
  RESIDENCE = "residence",
  STATUE = "statue",
  STORE = "store",
  TAVERN = "tavern",
  WAREHOUSE = "warehouse",
  WATCHTOWER = "watchtower",
  WELL = "well",
  WHEAT_FIELD = "wheat field",
  WORKSHOP = "workshop",
}